task1ver4

+++ modules/task_1/kasyanychev_m_min_elem/kasyanychev_m_min_elem.cpp
@@ -18,7 +18,7 @@ std::vector<int> RandomVector(int n) {
 
 int SequentialOperations(std::vector<int> x) {
     int min = x[0];
-    for (int i = 1; i < x.size(); i++) {
+    for (long unsigned int i = 1; i < x.size(); i++) {
         if (x[i] < min)
             min = x[i];
     }
